# crates/dagron-engine/openapi.yaml
openapi: 3.0.3
info:
  title: module-54 workflow scheduler — management API
  description: >
    Read/write control surface over the workflow scheduler's datastore: list and
    inspect runs, submit a DAG, cancel a run, and scrape metrics. Every operation
    is a thin shell over the same `db` facade the reconcile loop uses, so the API
    holds no scheduler state of its own and works whether it runs in-process next
    to the loop or as a standalone read sidecar against the shared Postgres
    backend.


    Served only when the binary is built with the (default) `ops` Cargo feature
    and `API_ADDR` is set. This document is also served live by the running
    scheduler at `/openapi.yaml` and `/openapi.json`, with a Swagger UI at `/docs`.


    Security: this embedded API performs **no authentication** of its own — bind
    `API_ADDR` to localhost / a trusted network and place authentication (e.g. the
    `dagron-api` service, an auth proxy, or network policy) in front of it before
    exposing it. It is intended for in-process ops and trusted-network sidecar use.
  version: 0.1.0
  license:
    name: lab crate
servers:
  - url: http://127.0.0.1:8080
    description: Local management API (set API_ADDR to choose host:port)
...
paths:
...
  /dead-letters/{id}/redrive:
    post:
      tags: [dead-letters]
      summary: Redrive a dead letter
      description: >
        Re-attempt the parked payload as a fresh run. On success the dead letter
        is deleted; a still-invalid payload returns `400` and the row is kept.
      operationId: redriveDeadLetter
      parameters:
        - $ref: "#/components/parameters/DeadLetterId"
      responses:
        "200":
          description: Redriven into a new run
          content:
            application/json:
              schema:
                type: object
                required: [run_id, redriven_from]
                properties:
                  run_id: { type: string, format: uuid }
                  redriven_from: { type: string, format: uuid }
        "400":
          description: The parked payload is still an invalid DAG
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: No such dead letter
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "500":
          $ref: "#/components/responses/Error"
  /dead-letters/{id}:
    delete:
      tags: [dead-letters]
      summary: Discard a dead letter
      operationId: deleteDeadLetter
      parameters:
        - $ref: "#/components/parameters/DeadLetterId"
      responses:
        "200":
          description: Discarded
          content:
            application/json:
              schema:
                type: object
                required: [id, deleted]
                properties:
                  id: { type: string, format: uuid }
                  deleted: { type: boolean, example: true }
        "404":
          description: No such dead letter
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "500":
          $ref: "#/components/responses/Error"
...
components:
  parameters:
...
    DeadLetterId:
      name: id
      in: path
      required: true
      description: Dead-letter UUID.
      schema: { type: string, format: uuid }
  responses:
    Error:
      description: Unexpected server/datastore error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string, description: Human-readable failure message }
...